requirement.txt
langchain
python-dotenv==1.0.0
streamlit=
protobuf~=3.19.0
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
chromadb
dotenv-python
langchain_experimental
langchain_google_genai
pymysql
google-generativeai